chore: more v1 test suite cleanups
